Remaud adder

+++ src/qrisp/alg_primitives/mcx_algs/khattar.py
@@ -59,7 +59,15 @@ def ctrl_state_conjugator(ctrls, ctrl_state):
             x(ctrls[i])
 
 @custom_inversion
-def gidney_CCCZ(ctrls, target, inv = False):
+def gidney_CCCZ(ctrls, target, inv=False):
+    """
+    Implements a CCCZ gate using the Gidney 
+    method described in https://arxiv.org/abs/2106.11513 using only 6 T gates.
+    Args:
+        ctrls (list): A list of control qubits. It is expected to contain three qubits.
+        target (list): A list containing the target qubit. It is expected to contain one qubit.
+    """
+    
     gidney_anc = QuantumFloat(1)
 
     h(gidney_anc[0])
@@ -76,6 +84,7 @@ def gidney_CCCZ(ctrls, target, inv = False):
     cx(ctrls[2], gidney_anc[0])
     t_dg(gidney_anc[0])
     cx(target[0], gidney_anc[0])
+    
     with invert():
         sx(gidney_anc[0])
     cl_res = measure(gidney_anc[0])
@@ -92,7 +101,7 @@ def gidney_CCCZ(ctrls, target, inv = False):
 
 
 def cca_4ctrls(ctrls, target):
-
+    #This function handles the case of 4 control qubits in the conditionally clean ancillae MCX
     cca4_anc = QuantumFloat(1)
 
     mcx([ctrls[1], ctrls[0]], cca4_anc[0], method="gray_pt")
@@ -116,6 +125,7 @@ def cca_mcx(ctrls, target, anc):
         xrange = range
     else:
         xrange = jrange
+        
     # STEP 1
     mcx([ctrls[0], ctrls[1]], anc[0], method="gidney")
 
@@ -141,6 +151,24 @@ def cca_mcx(ctrls, target, anc):
 
 @qache
 def khattar_mcx(ctrls, target, ctrl_state):
+    """
+    Implements the Khattar multi-controlled X (MCX) gate methode using conditionally clean ancillae described in https://arxiv.org/abs/2407.17966.
+    The behavior of the function varies depending on the number of 
+    control qubits (N) and the control state.
+    Args:
+        ctrls (list): A list of control qubits.
+        target (list): A list containing the target qubit(s).
+        ctrl_state (int or str): The control state, which can be provided as an integer or a binary string.
+            If provided as a string, it is reversed and converted to an integer.
+    Behavior:
+        - For N = 1: A single-controlled X gate is applied.
+        - For N = 2: A two-controlled X gate is applied.
+        - For N = 3: If in tracing mode, a decomposition using Hadamard gates and a CCCZ gate is applied.
+          Otherwise, a specific MCX method ("balauca") is used.
+        - For N = 4: A custom 4-controlled gate (`cca_4ctrls`) is applied.
+        - For N > 4: An ancillary qubit is used to decompose the operation into smaller steps.
+    """
+    
     N = jlen(ctrls)
 
     if isinstance(ctrl_state, str):
@@ -151,33 +179,66 @@ def khattar_mcx(ctrls, target, ctrl_state):
 
     with conjugate(ctrl_state_conjugator)(ctrls, ctrl_state):
 
-        with control(N == 1):
-            cx(ctrls[0], target[0])
+        if isinstance(ctrls, list):
+            if N == 1:
+                cx(ctrls[0], target[0])
 
-        with control(N == 2):
-            mcx([ctrls[0], ctrls[1]], target[0])
+            if N == 2:
+                mcx([ctrls[0], ctrls[1]], target[0])
 
-        with control(N == 3):
-            if check_for_tracing_mode():
-                h(target[0])
-                gidney_CCCZ(ctrls, target)
-                h(target[0])
-            else:
-                mcx(ctrls, target[0], method="balauca")  # CHANGE
+            if N == 3:
+                if check_for_tracing_mode():
+                    #Conjugating with H to transform the CCCZ into a CCCX
+                    h(target[0])
+                    gidney_CCCZ(ctrls, target)
+                    h(target[0])
+                else:
+                    mcx(ctrls, target[0], method="balauca")  # CHANGE
 
-        with control(N == 4):
-            cca_4ctrls(ctrls, target)
+            if N == 4:
+                cca_4ctrls(ctrls, target)
 
-        with control(N > 4):
-            khattar_anc = QuantumFloat(1)
-            with conjugate(cca_mcx)(ctrls, target, khattar_anc):
-                # STEP 3
-                mcx([khattar_anc[0], ctrls[0]], target[0])
-            khattar_anc.delete()
+            if N > 4:
+                khattar_anc = QuantumFloat(1)
+                with conjugate(cca_mcx)(ctrls, target, khattar_anc):
+                    # STEP 3
+                    mcx([khattar_anc[0], ctrls[0]], target[0])
+                khattar_anc.delete()
+
+        else:
+            with control(N == 1):
+                cx(ctrls[0], target[0])
+
+            with control(N == 2):
+                mcx([ctrls[0], ctrls[1]], target[0])
+
+            with control(N == 3):
+                if check_for_tracing_mode():
+                    #Conjugating with H to transform the CCCZ into a CCCX
+                    h(target[0])
+                    gidney_CCCZ(ctrls, target)
+                    h(target[0])
+                else:
+                    mcx(ctrls, target[0], method="balauca")  # CHANGE
+
+            with control(N == 4):
+                cca_4ctrls(ctrls, target)
+
+            with control(N > 4):
+                khattar_anc = QuantumFloat(1)
+                with conjugate(cca_mcx)(ctrls, target, khattar_anc):
+                    # STEP 3
+                    mcx([khattar_anc[0], ctrls[0]], target[0])
+                khattar_anc.delete()
 
 
 @qache
 def khattar_mcp(phi, ctrls, ctrl_state):
+    """
+    Implements the multi-controlled phase (MCP) gate based on the Khattar MCX implementation.
+    
+    """
+    
     N = jlen(ctrls)
 
     if isinstance(ctrl_state, str):
@@ -201,8 +262,8 @@ def khattar_mcp(phi, ctrls, ctrl_state):
                 h(target[0])
                 gidney_CCCZ(ctrls, target)
                 h(target[0])
-
-                p(phi, target[0])  # Use a conjugation environment also here? Maybe yes
+                
+                p(phi, target[0])
 
                 h(target[0])
                 gidney_CCCZ(ctrls, target)
